Ease writing of new on-disk Binding implementations
